+++ src/BizHawk.Client.Common/lua/LuaHelperLibs/BitLuaLibrary.cs
@@ -70,7 +70,7 @@ namespace BizHawk.Client.Common
 		[LuaMethod("rshift", "Logical shift right of 'val' by 'amt' bits")]
 		public static uint Rshift(uint val, int amt)
 		{
-			return (uint)(val >> amt);
+			return val >> amt;
 		}
 
 		[LuaMethodExample("local inbitars = bit.arshift( -1000, 4 );")]
@@ -91,7 +91,7 @@ namespace BizHawk.Client.Common
 		[LuaMethod("set", "Sets the bit 'pos' in 'num'")]
 		public static uint Set(uint num, int pos)
 		{
-			return (uint)(num | 1U << pos);
+			return num | 1U << pos;
 		}
 
 		[LuaMethodExample("local lobitcle = bit.clear( 25, 35 );")]
